+++ src/Adapter/DbTable/CallbackCheckAdapter.php
@@ -1,18 +1,17 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-authentication for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-authentication/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-authentication/blob/master/LICENSE.md New BSD License
  */
 
-namespace Zend\Authentication\Adapter\DbTable;
+namespace Laminas\Authentication\Adapter\DbTable;
 
-use Zend\Authentication\Result as AuthenticationResult;
-use Zend\Db\Adapter\Adapter as DbAdapter;
-use Zend\Db\Sql;
-use Zend\Db\Sql\Predicate\Operator as SqlOp;
+use Laminas\Authentication\Result as AuthenticationResult;
+use Laminas\Db\Adapter\Adapter as DbAdapter;
+use Laminas\Db\Sql;
+use Laminas\Db\Sql\Predicate\Operator as SqlOp;
 
 class CallbackCheckAdapter extends AbstractAdapter
 {
@@ -27,20 +26,20 @@ class CallbackCheckAdapter extends AbstractAdapter
     /**
      * __construct() - Sets configuration options
      *
-     * @param DbAdapter $zendDb
+     * @param DbAdapter $laminasDb
      * @param string    $tableName                    Optional
      * @param string    $identityColumn               Optional
      * @param string    $credentialColumn             Optional
      * @param callable  $credentialValidationCallback Optional
      */
     public function __construct(
-        DbAdapter $zendDb,
+        DbAdapter $laminasDb,
         $tableName = null,
         $identityColumn = null,
         $credentialColumn = null,
         $credentialValidationCallback = null
     ) {
-        parent::__construct($zendDb, $tableName, $identityColumn, $credentialColumn);
+        parent::__construct($laminasDb, $tableName, $identityColumn, $credentialColumn);
 
         if (null !== $credentialValidationCallback) {
             $this->setCredentialValidationCallback($credentialValidationCallback);
@@ -69,7 +68,7 @@ class CallbackCheckAdapter extends AbstractAdapter
     }
 
     /**
-     * _authenticateCreateSelect() - This method creates a Zend\Db\Sql\Select object that
+     * _authenticateCreateSelect() - This method creates a Laminas\Db\Sql\Select object that
      * is completely configured to be queried against the database.
      *
      * @return Sql\Select