+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#4](https://github.com/zendframework/zend-authentication/pull/4) adds
+- [zendframework/zend-authentication#4](https://github.com/zendframework/zend-authentication/pull/4) adds
   documentation, which can be compiled using [bookdown](http://bookdown.io):
   `bookdown doc/bookdown.json`; docs can then be viewed by starting a web server
   via `php -S 0.0.0.0:8000 -t doc/html/` and browsing to http://localhost:8000/.